requirements.txt
tsdat >=0.3.1, <0.4.0
cmocean
notebook
act-atmos
